feat: refactor SendFundsToCollectiveSection.js as a new component
