+++ components/edit-collective/actions/EmptyBalance.js
@@ -1,91 +1,18 @@
-import React, { Fragment, useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import { CollectiveType } from '../../../lib/constants/collectives';
-import { formatCurrency } from '../../../lib/currency-utils';
 
 import Container from '../../Container';
-import SendMoneyToCollectiveBtn from '../../SendMoneyToCollectiveBtn';
-import StyledButton from '../../StyledButton';
-import Modal, { ModalBody, ModalFooter, ModalHeader } from '../../StyledModal';
 import { P } from '../../Text';
 import SettingsSectionTitle from '../sections/SettingsSectionTitle';
+import SendFundsToCollectiveSection from '../SendFundsToCollectiveSection';
 
 const EmptyBalance = ({ collective, LoggedInUser }) => {
-  const [modal, setModal] = useState({ type: 'Transfer', show: false, isApproved: false });
-
   if (!collective.host || collective.host.id === collective.id) {
     return null;
   }
-
-  const confirmTransfer = () => {
-    setModal({ ...modal, show: true, isApproved: false });
-  };
-
-  const closeModal = () => setModal({ ...modal, show: false, isApproved: false });
-
-  const renderSendMoneyToCollectiveSection = toCollective => {
-    return (
-      <Fragment>
-        {collective.stats.balance > 0 && (
-          <SendMoneyToCollectiveBtn
-            fromCollective={collective}
-            toCollective={toCollective}
-            LoggedInUser={LoggedInUser}
-            amount={collective.stats.balance}
-            currency={collective.currency}
-            confirmTransfer={confirmTransfer}
-            isTransferApproved={modal.isApproved}
-          />
-        )}
-        {collective.stats.balance === 0 && (
-          <StyledButton disabled={true}>
-            <FormattedMessage
-              id="SendMoneyToCollective.btn"
-              defaultMessage="Send {amount} to {collective}"
-              values={{
-                amount: formatCurrency(0, collective.currency),
-                collective: toCollective.name,
-              }}
-            />
-          </StyledButton>
-        )}
-        <Modal show={modal.show} width="570px" onClose={closeModal}>
-          <ModalHeader onClose={closeModal}>
-            <FormattedMessage
-              id="collective.emptyBalance.header"
-              values={{ action: modal.type }}
-              defaultMessage={'{action} Balance'}
-            />
-          </ModalHeader>
-          <ModalBody>
-            <P>
-              <FormattedMessage
-                id="collective.emptyBalance.body"
-                values={{ collective: toCollective.name, action: modal.type.toLowerCase() }}
-                defaultMessage={'Are you sure you want to {action} to {collective}?'}
-              />
-            </P>
-          </ModalBody>
-          <ModalFooter>
-            <Container display="flex" justifyContent="flex-end">
-              <StyledButton mx={20} onClick={() => setModal({ ...modal, show: false, isApproved: false })}>
-                <FormattedMessage id="actions.cancel" defaultMessage={'Cancel'} />
-              </StyledButton>
-              <StyledButton
-                buttonStyle="primary"
-                data-cy="action"
-                onClick={() => setModal({ ...modal, show: false, isApproved: true })}
-              >
-                <FormattedMessage id="confirm" defaultMessage={'Confirm'} />
-              </StyledButton>
-            </Container>
-          </ModalFooter>
-        </Modal>
-      </Fragment>
-    );
-  };
 
   return (
     <Container display="flex" flexDirection="column" width={1} alignItems="flex-start" mb={50}>
@@ -117,12 +44,20 @@ const EmptyBalance = ({ collective, LoggedInUser }) => {
           />
         </P>
       )}
-      {[CollectiveType.FUND, CollectiveType.COLLECTIVE].includes(collective.type) &&
-        collective.host.hostCollective &&
-        renderSendMoneyToCollectiveSection(collective.host.hostCollective)}
-      {[CollectiveType.PROJECT, CollectiveType.EVENT].includes(collective.type) &&
-        collective.parentCollective &&
-        renderSendMoneyToCollectiveSection(collective.parentCollective)}
+      {[CollectiveType.FUND, CollectiveType.COLLECTIVE].includes(collective.type) && collective.host.hostCollective && (
+        <SendFundsToCollectiveSection
+          LoggedInUser={LoggedInUser}
+          collective={collective}
+          toCollective={collective.host.hostCollective}
+        />
+      )}
+      {[CollectiveType.PROJECT, CollectiveType.EVENT].includes(collective.type) && collective.parentCollective && (
+        <SendFundsToCollectiveSection
+          LoggedInUser={LoggedInUser}
+          collective={collective}
+          toCollective={collective.parentCollective}
+        />
+      )}
     </Container>
   );
 };
